Clickable Hierarchy and Example Scrollbar Container It's currently not the prettiest, but there is now a clickable hierarchy that can be used to both render the entire GUI and to propagate clicks into the GUI to the appropriate object. There is also, to test this, in the src/containers folder defined a Scrollbar_Container object with a constructor function that is used to test the hierarchy with.
The testing file is Clickable_Hierarchy.c in the examples folder. As seeable from all the commented out code, the hierarchy has the potential to really simplify the rendering and the clicking functionalities of a GUI setup.

The next step for the hierarchy system would be to optimize it. Another important thing to do would be to figure out the whole container thing. It's a good question whether it should even be a part of the core source or simply a recommended methodology. It's difficult to come up with a container setup that is generally useful enough to be able to justify putting into the core library. A sublibrary with various pre-built solutions is an idea to consider.